modulo de transações com operacao de transferencia

+++ modules/User/database/migrations/2025_01_13_133121_create_user_accounts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('user_accounts', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->decimal('amount', 10, 2)->default(0);
+            $table->bigInteger('amount')->default(0);
             $table->enum('type', [1, 2])->default(1);
             $table->foreign('user_id')->references('id')->on('users');
         });
